// src/app/interfaces/requestInterface/meeting.ts
export interface Meeting {
  // meetingID: number;
  title: string;
  description: string;
  meetingType: string;
  meetingDate: string;
  //branch: string;
  hostBy: string;
  videoEmbedCode: string;
  chatEmbedCode: string;
  branchID: number;
  publish: boolean;
  // dateCreated: number;
  // dateUpdated: number;
}